project1_rag/requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-anthropic
sentence-transformers
chromadb
pymupdf
markdown
python-dotenv
tiktoken
